requirements.txt
supabase==1.0.0
requests
aiohttp
python-dotenv
tqdm
